feat: make WarmStorage approvals infinite, focus only on deposit (#47)
Remove all of the complexity of dealing with approvals and rate limits by just
setting them to be infinite by default for WarmStorage (i.e. it's a trusted
service and your risk is managed by deposit). This lets us trim down the
UI for dealing with payments and simplifies the messaging.