feat: add data-set command (#50)
* feat: make WarmStorage approvals infinite, focus only on deposit

Remove all of the complexity of dealing with approvals and rate limits by just
setting them to be infinite by default for WarmStorage (i.e. it's a trusted
service and your risk is managed by deposit). This lets us trim down the
UI for dealing with payments and simplifies the messaging.

* feat: add data-set command

* fix: only display data-sets created by filecoin-pin

* feat: list by default if no datasetId given

* chore: fix lint

* chore: normalize data-set command to match others

* test: add basic data-set tests

* fix: prevent duplicate dataset list with [id] --ls

* feat: more service details for dataset

* chore: remove pointless Object.create(null)

---------

Co-authored-by: Rod Vagg <rod@vagg.org>